# com.trekglobal.idempiere.rest.api/openapi/idempiere-rest.yml
openapi: 3.0.0
info:
  title: iDempiere REST endpoints
  description: >-
    iDempiere REST endpoints.
  version: 1.0.0
servers:
  - url: '{base_url}'
    variables:
      base_url:
        enum:
          - 'http://localhost:8080/api/v1'
          - 'https://localhost:8443/api/v1'
          - 'https://test.idempiere.org/api/v1'
        default: 'http://localhost:8080/api/v1'
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes  
...
  responses:
...
    ServerError:
      description: Internal Server Error. Unhandled exceptions.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internal_server_error:
              value:
                title: "Internal Server Error"
                status: 500
                detail: "Internal server error. Unhandled exceptions."
...
  schemas:
    AuthenticationParameters:
      description: authentication parameters
      type: object
      properties:
        clientId:
          type: string
          description: Tenant id or Tenant search key
          example: 11 or GardenWorld
        roleId:
          type: string
          description: Role id or Role name
          example: 102 or GardenWorld Admin
        organizationId:
          type: string
          description: Organization id or Organization name
          example: 11 or HQ
        warehouseId:
          type: string
          description: Warehouse identifier
          example: 103 or HQ Warehouse
        language:
          type: string
          description: Language code
          example: en_US
    AuthenticationRequest:
      type: object
      properties:
        userName:
          type: string
          description: Username for authentication
        password:
          type: string
          description: Password for authentication
        parameters:
          $ref: '#/components/schemas/AuthenticationParameters'
    AuthenticationResponse:
      description: authentication response
      type: object
      properties:
         clients:
           description: clients accessible.
           type: array
           items:
             properties:
               id:
                 type: integer
               name:
                 type: string
         userId:
            type: integer
            description: User identifier
            example: 100
         language:
            type: string
            description: Language code (e.g., en_US)
            example: en_US
         menuTreeId:
            type: integer
            description: Menu tree identifier
            example: 10
         token:
            type: string
            description: Authentication token (JWT)
            example: "eyJraWQiOiJpZ..."
         refresh_token:
            type: string
            description: Refresh token (JWT)
            example: "eyJraWQiOiJpZ..."
    JWKS:
      type: object
      properties:
        keys:
          type: array
          items:
            type: object
            properties:
              alg:
                type: string
                description: Algorithm used for the key.
              k:
                type: string
                description: The key value (base64url encoded for 'oct' key type).
              kid:
                type: string
                description: Key ID.
              kty:
                type: string
                description: Key type.
                enum: # Recommended to include possible values
                  - oct
                  - RSA # Add other possible values as needed
                  - EC
      description: The JSON Web Key Set (JWKS) is a set of keys containing the public keys used to verify JSON Web Token (JWT)
    ErrorResponse:
      description: Error response from server
      type: object
      properties:
        title:
          type: string
          description: error summary
        status:
          type: integer
          description: http status code
        detail:
          type: string
          description: error details
...
security:
    - bearerAuth: [] # use the same name as above

# The x-* below is for reference only    
...
paths:
  /auth/tokens:
    post:
      tags:
        - Authentication
      summary: Login for authorization token
      description: >-
        Login with tenant, user name, password and role. The return
        token is use as Authorization request header (Bearer {token}) for
        subsequent api call

      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationRequest'              
            examples:
              gardenAdminUserNameWithoutParameters:
                summary: GardenAdmin UserName authentication without parameters. Follow with PUT api/v1/auth/tokens to set login parameters.
                value:
                  userName: "GardenAdmin"
                  password: "GardenAdmin"
              gardenAdminEmailWithoutParameters:
                summary: GardenAdmin Email authentication without parameters. Follow with PUT api/v1/auth/tokens to set login parameters.
                value:
                  userName: "admin @ gardenworld.com"
                  password: "GardenAdmin"
              gardenAdminUserNameOneStepLogin:
                summary: GardenAdmin UserName authentication with integer id login parameters.
                value:
                  userName: "GardenAdmin"
                  password: "GardenAdmin"
                  parameters:
                    clientId: 11
                    roleId: 102
                    organizationId: 11
                    warehouseId: 103
                    language: en_US
              gardenAdminEmailOneStepLogin:
                summary: GardenAdmin Email authentication with integer id login parameters.
                value:
                  userName: "admin @ gardenworld.com"
                  password: "GardenAdmin"
                  parameters:
                    clientId: 11
                    roleId: 102
                    organizationId: 11
                    warehouseId: 103
                    language: en_US
              gardenAdminUserNameOneStepLoginWithStringIdentifier:
                summary: GardenAdmin UserName authentication with string identifier login parameters.
                value:
                  userName: "GardenAdmin"
                  password: "GardenAdmin"
                  parameters:
                    clientId: GardenWorld
                    roleId: GardenWorld Admin
                    organizationId: HQ
                    warehouseId: HQ Warehouse
                    language: en_US
              gardenAdminEmailOneStepLoginWithStringIdentifier:
                summary: GardenAdmin Email authentication with string identifier login parameters.
                value:
                  userName: "admin @ gardenworld.com"
                  password: "GardenAdmin"
                  parameters:
                    clientId: GardenWorld
                    roleId: GardenWorld Admin
                    organizationId: HQ
                    warehouseId: HQ Warehouse
                    language: en_US
              gardenAdminUserNameOneStepLoginWithDefaults:
                summary: GardenAdmin UserName authentication with partial login parameters.
                value:
                  userName: "GardenAdmin"
                  password: "GardenAdmin"
                  parameters:
                    roleId: GardenWorld Admin
                    organizationId: HQ
                    warehouseId: HQ Warehouse
              gardenAdminEmailOneStepLoginWithDefaults:
                summary: GardenAdmin Email authentication with partial login parameters.
                value:
                  userName: "admin @ gardenworld.com"
                  password: "GardenAdmin"
                  parameters:
                    roleId: GardenWorld Admin
                    organizationId: HQ
                    warehouseId: HQ Warehouse
              superUserUserNameSystemTenant:
                summary: SuperUser UserName authentication for System tenant.
                value:
                  userName: "SuperUser"
                  password: "System"
                  parameters:
                    clientId: 0
                    roleId: 0 #Need to change System Administrator role type to null
                    organizationId: 0
                    language: en_US
              superUserEmailSystemTenant:
                summary: SuperUser Email authentication for System tenant.
                value:
                  userName: "superuser @ idempiere.com"
                  password: "System"
                  parameters:
                    clientId: 0
                    roleId: 0 #Need to change System Administrator role type to null
                    organizationId: 0
                    language: en_US
      responses:
        '200':
          description: OK. Credentials are valid; returns client list or full session token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
              examples:
                withAuthenticationParameters:
                  value:
                    userId: 100
                    language: en_US
                    menuTreeId: 10
                    token: "eyJraWQiOiJpZ..."
                    refresh_token: "eyJraWQiOiJpZ..."
                withoutAuthenticationParameters:
                  value:
                    clients:
                      - id: 11
                        name: GardenWorld
                    token: "eyJraWQiOiJpZ..."
        '400':
          description: Bad Request. Error during JWT creation or invalid data format in parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized. Wrong password, account locked, or no access to the requested Client/Role/Org.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidUserNameOrPassword:
                  value:
                    title: 'Authenticate error'
                    status: 401
                    detail: 'Invalid User ID or Password'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - Authentication
      summary: Set login parameters
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationParameters'
            examples:
              withId:
                summary: integer id login parameters.
                value:
                  clientId: 11
                  roleId: 102
                  organizationId: 11
                  warehouseId: 103
                  language: en_US
              withStringIdentifier:
                summary: string identifier login parameters.
                value:
                  clientId: GardenWorld
                  roleId: GardenWorld Admin
                  organizationId: HQ
                  warehouseId: HQ Warehouse
                  language: en_US
      responses:
        '200':
          description: OK. Session context successfully updated; new token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
              example:
                userId: 100
                language: en_US
                menuTreeId: 10
                token: "eyJraWQiOiJpZ..."
                refresh_token: "eyJraWQiOiJpZ..."
        '400':
          description: Bad Request. Technical error during token generation or illegal arguments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized. Permission denied for the requested Client/Role/Org, or parameters are missing/ambiguous.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidUserNameOrPassword:
                  value:
                    title: 'Authenticate error'
                    status: 401
                    detail: 'Invalid User ID or Password'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh authorization token
      description: >-
        The authorization token expires in 1 hour and the refresh_token expires in 24 hours.
        After the authorization token expires, you can get a new authorization token using the refresh token.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: 
                  type: string
                  description: Refresh token from /auth/tokens or /auth/refresh call
                clientId:
                  type: integer
                  description: clientId from the source /auth/tokens or /auth/refresh call
                userId:
                  type: integer
                  description: userId from the /auth/tokens response
            example:
              refresh_token: 'eyJraWQiOiJpZGVtcG...'
              clientId: 11
              userId: 101
      responses:
        '200':
          description: OK. Token successfully rotated; returns new auth and refresh tokens.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                      type: string
                      description: Authentication token (JWT)
                      example: "eyJraWQiOiJpZ..."
                  refresh_token:
                      type: string
                      description: Refresh token (JWT)
                      example: "eyJraWQiOiJpZ..."
              example:
                token: "eyJraWQiOiJpZ..."
                refresh_token: "eyJraWQiOiJpZ..."
        '400':
          description: Bad Request. Technical error during the signing of the new JWT.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized. Token is invalid, expired, already used (reuse detection), or ownership verification failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidUserNameOrPassword:
                  value:
                    title: 'Authenticate error'
                    status: 401
                    detail: 'Invalid User ID or Password'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout the session related to a token.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token: 
                  type: string
                  description: token from /auth/tokens or /auth/refresh call
      responses:
        '200':
          description: OK. Session successfully terminated and tokens invalidated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                    description: info or error message from server
        '401':
          description: Unauthorized. Token signature is invalid or malformed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidUserNameOrPassword:
                  value:
                    title: 'Authenticate error'
                    status: 401
                    detail: 'Invalid User ID or Password'
        '404':
          description: Not Found. Token is valid but missing a session ID (not an active session token).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/password-reset/request:
    post:
      tags:
        - Authentication
      summary: Password reset - request a one-time code
      description: >-
        First step of the code-based password reset. Emails a one-time numeric code to
        the address. The response is always neutral: it never reveals whether the email
        matches an account. Unauthenticated - no Authorization header required.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  description: Email / login of the account to reset
                language:
                  type: string
                  description: Optional AD_Language for the email template (e.g. en_US)
            example:
              email: user@example.com
              language: en_US
      responses:
        '200':
          description: OK. Neutral confirmation, returned whether or not the account exists.
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                    description: Neutral confirmation message
        '429':
          description: Too Many Requests. The per-identifier request rate limit was exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/password-reset/verify:
    post:
      tags:
        - Authentication
      summary: Password reset - verify the code
      description: >-
        Second step of the code-based password reset. Verifies the emailed code and, on
        success, returns a short-lived single-use verified token. Unauthenticated.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - code
              properties:
                email:
                  type: string
                code:
                  type: string
                  description: The one-time code emailed to the address
            example:
              email: user@example.com
              code: "123456"
      responses:
        '200':
          description: OK. Code verified.
          content:
            application/json:
              schema:
                type: object
                properties:
                  verifiedToken:
                    type: string
                    description: Short-lived single-use token for /auth/password-reset/complete
        '400':
          description: >-
            Bad Request. Invalid code, expired code, or too many attempts. The response is
            uniform and does not reveal whether the email is registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/password-reset/complete:
    post:
      tags:
        - Authentication
      summary: Password reset - set the new password
      description: >-
        Third step of the code-based password reset. Uses the verified token from the
        verify step to set the new password; existing sessions for the account are
        invalidated. Unauthenticated.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - verifiedToken
                - newPassword
              properties:
                verifiedToken:
                  type: string
                  description: Token returned by /auth/password-reset/verify
                newPassword:
                  type: string
            example:
              verifiedToken: 3f9a0b7c...
              newPassword: MyNewPassw0rd!
      responses:
        '200':
          description: OK. Password changed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
        '400':
          description: >-
            Bad Request. Invalid or expired verified token, or the new password violates
            the password policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/jwk:
    get:
      tags:
        - Authentication
      summary: Get the signing key for integration with gateway API. 
      description: This requires the SysConfig IDEMPIERE_REST_EXPOSE_JWK=Y.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWKS'
        '400':
          description: Bad Request. Configuration error preventing the algorithm initialization.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/roles:
    get:
      tags:
        - Authentication
      summary: Get list of roles avalable
      parameters:
        - name: client
          in: query
          required: true
          schema:
            type: integer
          description: client id
      responses:
        '200':
          description: OK. Returns the list of authorized roles for the specified client.
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: The unique identifier for the role.
                        name:
                          type: string
                          description: The name of the role.
                    description: An array of roles.
                description: A collection of roles.
        '400':
          description: Bad Request. Illegal arguments or error in retrieving client/role information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized. The user does not have access to the requested clientId.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidUserNameOrPassword:
                  value:
                    title: 'Authenticate error'
                    status: 401
                    detail: 'Invalid User ID or Password'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/organizations:
    get:
      tags:
        - Authentication
      summary: Get list of organizations avalable
      parameters:
        - name: client
          in: query
          required: true
          schema:
            type: integer
          description: client id
        - name: role
          in: query
          required: true
          schema:
            type: integer
          description: role id
      responses:
        '200':
          description: OK. Returns the list of authorized organizations for the specified client and role.
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: The unique identifier for the organization.
                        name:
                          type: string
                          description: The name of the organization.
                    description: An array of organization.
                description: A collection of organization.
        '400':
          description: Bad Request. Illegal arguments or error in retrieving Client/User/Role information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized. The user does not have access to the requested clientId or roleId.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidUserNameOrPassword:
                  value:
                    title: 'Authenticate error'
                    status: 401
                    detail: 'Invalid User ID or Password'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/warehouses:
    get:
      tags:
        - Authentication
      summary: Get list of warehouse avalable
      parameters:
        - name: client
          in: query
          required: true
          schema:
            type: integer
          description: client id
        - name: role
          in: query
          required: true
          schema:
            type: integer
          description: role id
        - name: organization
          in: query
          required: true
          schema:
            type: integer
          description: organization id
      responses:
        '200':
          description: OK. Returns the list of authorized warehouses for the specified context.
          content:
            application/json:
              schema:
                type: object
                properties:
                  warehouses:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: The unique identifier for the warehouse.
                        name:
                          type: string
                          description: The name of the warehouse.
                    description: An array of warehouse.
                description: A collection of warehouse.
        '400':
          description: Bad Request. Illegal arguments or error in retrieving context information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized. The user does not have access to the requested clientId, roleId, or organizationId.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidUserNameOrPassword:
                  value:
                    title: 'Authenticate error'
                    status: 401
                    detail: 'Invalid User ID or Password'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/language:
    get:
      tags:
        - Authentication
      summary: Get default client language
      parameters:
        - name: client
          in: query
          required: true
          schema:
            type: integer
          description: client id
      responses:
        '200':
          description: OK. Returns the default language code for the specified client.
          content:
            application/json:
              schema:
                type: object
                properties:
                  AD_Language:
                    type: string
                    description: language code
                    example: en_US
        '400':
          description: Bad Request. Invalid clientId or internal mapping error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
...